Bugfix: calling update_flexilims was not changing the `not_online` status
If a dataset name had no `_index` at the end, it was added by `add_dataset` to
the flexilims name but the dataset instance was not updated. A second call to
update_flexilims was therefore not seing the new dataset and creating yet another
one with the next index (as add_dataset is `append` by default)

Fix:
 Update ds.name to match the new index
 Call add_dataset in `abort` mode as it should be called only when the dataset
 is not already on flexilims.

At least if there was already one online with this name. Now after the first call,
the name of the dataset is updated to match the new generated name.