dataset.flexilims_report look of uploaded differences
Dataset.flexilims_report looks for differences between current dataset state and
values read from flexilims. However, at upload, flexilims replaces empty structures
('', {}, [] ...) by null, which is downloaded as a None. Therefore you can have a
difference that is still present after updating flexilims.
To correct for that ignore empty structures that match to a None
element on flexilims in the report but print a warning.

Update test_datasets.py to reflect this change